requirements.txt
requests
matplotlib
pillow

scikit-learn
numpy

tqdm

black

# packet_analysis.py
pyshark
levenshtein